pyproject.toml
[project]
name = "open-data-platform"
version = "0.1.0"
description = "Open Data Platform with Fabric-compatible local development"
requires-python = ">=3.9"
dependencies = [
    "pyspark>=3.5.0",
    "delta-spark>=3.0.0",
    "pandas>=2.0.0",
    "requests>=2.31.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pytz>=2024.1",
    "beautifulsoup4>=4.12.0",
    "dbt-core>=1.7.0",
    "dbt-postgres>=1.7.0",
    "acryl-datahub>=0.12.0",
    "acryl-datahub-airflow-plugin>=0.12.0",
    "boto3>=1.34.0",
    "psycopg2-binary>=2.9.0",
    "apache-airflow>=2.8.0",
    "apache-airflow-providers-apache-spark>=4.7.0",
    "ratelimit>=2.2.1",
    "opentelemetry-api>=1.23.0",
    "opentelemetry-sdk>=1.23.0",
    "opentelemetry-exporter-otlp>=1.23.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.0.0",
    "playwright>=1.50.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["shared*", "pipelines*", "scripts*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
markers = [
    "critical: Must-pass E2E critical path checks.",
    "e2e: Standard E2E checks.",
]

[tool.ruff]
line-length = 120
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_ignores = true
